test: fail test on error
It's very hard to have confidence in the tests as many continue to work
even when there's an error. By failing on error, we may get less data
but it will be more useful and better for building confidence. There may
need to be further adjustments to disable this for specific operations.

[#178142486](https://www.pivotaltracker.com/story/show/178142486)