fix: remove duplicate plan ID (part 2)
Previous fix #34 was sufficient to allow the brokerpak to install on a
fresh TAS 2.11 system, but upgrades from previous versions of Cloud Foundry
failed because there was still a reference to the duplicated ID in the
Cloud Foundry database. By removing all references to
`a556d1b4-5825-11ea-adb8-00155d4dfe6c`, we make sure that this error
cannot occur, although it does mean that service instances of two plans
will need to be updated to the *-v2 version of the plan.

[#178330774](https://www.pivotaltracker.com/story/show/178330774)