test: tests should work on CF 8, and earlier
CF CLI v8 will introduce some useful changes for services, specifically
the --wait flag for creating and deleting service instances. It also
introduces some incidental features such as a subtle difference in
output for the "cf service-key" command. This change allows the tests to
run with CF 8 and use the --wait flag, or to run on earlier versions and
still work