delete existing logins on unbind and ignore errors
we now create users, no logins. We ignore errors so if we get an error
when trying to delete a login that doesn't exist on unbind, it continues and drops
the user and all other dependencies.
This is to avoid leaving leftover logins in the db now that we moved to
create users instead.

[#179168006](https://www.pivotaltracker.com/story/show/179168006)