test: update examples to no longer use MySQL 5.6
When we try to create a MySQL 5.6 database we get:
```
Code="InvalidVersion" Message="Version '5.6' is not supported."
```

So we should update the examples test. It's probably just testing a
non-default version, so updated to test v8. Decided not to remove v5.6
from the allowed version enum because this might cause issues for people
who already have a v5.6 DB and are doing an update-service.

[#179433132](https://www.pivotaltracker.com/story/show/179433132)