fix: [azure-msssql-db-failover] Remove server_credentials from user_input
Providing server credentials as user inputs during provision and not bind causes binding to fail. It is unclear why this flow was provided so we are removing the option to supply the service credentials on provision time entirely.
`GSB_SERVICE_CSB_AZURE_MSSQL_DB_FAILOVER_GROUP_PROVISION_DEFAULTS` does not need to be specified when `MSSQL_DB_FOG_SERVER_PAIR_CREDS` env var is set.

[#179298527](https://www.pivotaltracker.com/story/show/179298527)