test(mongodb): use smaller plans
The test was updated to use medium->large plan change to rule out a
potential issue with the small plan. But we now now that the
small->medium plan change works just as well, so we should use that as
the main test since it may be faster and use less resources.

[#180197227](https://www.pivotaltracker.com/story/show/180197227)