test: get should retry
We have seen issues on MongoDB where we put some data with one app, then
fail to get it with a second app. We think this may be an issue with
eventual consistency in MongoDB, so as an experiment have added some
retry logic to get HTTP GET step so that it will keep trying until it
gets an HTTP 200 OK or times out after a minute.

[#181126724](https://www.pivotaltracker.com/story/show/181126724)