chore: fix csbsqlserver provider issue with dropping logins
We were able to query for existing logins using the Docker fake, but
this did not work in real Azure databases. Rather than go to reasonable
effort to fix this - and still have something potentially fragile, it
was decided to just attempt to drop the login and ignore any errors.
This matches the functionality that this replaces.

[#182571088](https://www.pivotaltracker.com/story/show/182571088)