fix(csbsqlserver): issue with validating server password
A test failed because a server had been created with a random password
containing an exclamation mark (!), which highlighted an issue where we
were using the same validation for the server password as for the
binding password. We can be opinionated as to characters allowed in the
binding password because this provider creates those users. But this
provider should not be opinionated as to the value of identifiers that
are passed in as server config. Rather we should do very basic validation
(check that there is a value) and rely on a URL escape to protect against
injection attacks.

[#182596458](https://www.pivotaltracker.com/story/show/182596458)