fix: amend run-failover tf files for tf > 0.12 (#267)
Previously the terraform files relied on variables to be interpolated
at the destroy step. This is no longer supported, instead the variables
required have been moved into a trigger block. In order to ensure the action
is performed when tf provision/destroy is called we added a uuid() to be generated
and ensure there is a trigger to allow the action.

[#182608274](https://www.pivotaltracker.com/story/show/182608274)