fix: build brokerpak on Mac
A well-intentioned optimisation has broken the ability to build a
brokerpak on a Mac. This is because on a Mac it's necessary to
cross-compile linux binaries, and Mac binaries are often useful too for
being able to run locally. We should probably avoid building Mac binaries
on a Linux desktop, and that could be a future enhancement.