fix: restore azurerm@v2.33 and v2.20 Terraform providers
- In previous versions, a `~>` provider version constraint was used
which restricts the version to patch fixes only.
- In this version we will update all the HCL with this constraint, but
just before the upgrade process we will perform a "terraform show" which
will evaluate the `~>` constraint and will fail if the matching version
of the provider is not present.
- Versions 2.33.0 and 2.20.0 have been shipped previously, and this
change adds both of those versions back until we are sure that all
customers have upgraded and the `~>` constraint is no longer in use.

[#182413672](https://www.pivotaltracker.com/story/show/182413672)