test: csb-azure-mssql-db-failover-group existing upgrade
An issue was found when upgrading the
"csb-azure-mssql-db-failover-group" service with the "existing" plan.
This is because the "existing" plan connects to an existing FOG, and
therefore expectations about what's in the Terraform state file don't
match. This test replicates the failure so that the upgrade can be
tested in future and the fix can be validated.

[#183463447](https://www.pivotaltracker.com/story/show/183463447)