chore(revert): mark sensitive values in var inputs and outputs (#356) (#361)
* Revert "chore: apply and enforce "terraform fmt" (#360)"

This reverts commit 0bf20c91fc75e9d13aa80a6ed5aba602df9d6e33.
It's necessary because we want to revert the previous commit, and this
changes the same files.

* Revert "feat: mark sensitive values in var inputs and outputs (#356)"

This reverts commit 728409acdfd08793b55a16d8bd01d9fc0fbd75db.

Revert as sensitive input variables are only possible with tf version >= `0.14`. Since the brokerpak possibly has service instances created with tf 0.13 or earlier, that blocks this feature because it would be a breaking change.