chore(revert): mark sensitive values in var inputs and outputs (#356) (#361)
* Revert "chore: apply and enforce "terraform fmt" (#360)"

This reverts commit c768baca6ba8be51e7c14b57a93341f8c366ae6e.
It's necessary because we want to revert the previous commit, and this
changes the same files.

* Revert "feat: mark sensitive values in var inputs and outputs (#356)"

This reverts commit cf91c7447093a53ecd980fa95436bf74f155fb3e.

Revert as sensitive input variables are only possible with tf version >= `0.14`. Since the brokerpak possibly has service instances created with tf 0.13 or earlier, that blocks this feature because it would be a breaking change.