chore: ensure that custom providers are statically linked
The custom Terraform providers should be statically linked. It looks
like a change in Go 1.20 resulted in them being dynamically linked when
built on Linux, and then they didn't work in the CSB app because the
binary buildpack doesn't have the required libraries available. This
change ensures that they are statically linked.

[#184404359](https://www.pivotaltracker.com/story/show/184404359)