fix: Redis upgrade from brokerpak 1.4.0 (#439)
* fix: Redis upgrade from brokerpak 1.4.0

- Prior to 1.4.0 there was no "redis_version" property.
- In 1.4.0 the redis_version property was added, and existing plans were
changed to specify redis_version=4. This was sensible at the time since
previous versions created Redis v4 instance.
- Recently a test started to fail because new Redis instances created
with brokerpak versions prior to 1.4.0 started to default to Redis v6
instances. This meant that when trying to upgrade to 1.4.0 or higher,
an attempt was made to change the Redis version from v6 -> v4,
resulting in the error:

  properties.redisVersion cannot be updated to 4.

- To resolve this, the value of redis_version for existing plans has been
removed so that the Azure default will be used. This allows the upgrade
to succeed for Redis v6.

[#184460134](https://www.pivotaltracker.com/story/show/184460134)

* fix: redis version is not expected

[#184460134](https://www.pivotaltracker.com/story/show/184460134)

---------

Co-authored-by: Andrea Zucchini <zandrea@vmware.com>