test: replace gorilla/mux with chi in test apps
gorilla/mux has been archived and subsequently removed from most test apps. Some test apps are sufficiently complex that a framework was needed in addition to the standard "net/http" package.

chi was chosen because of it plays well with "net/http" and has been the replacement for gorilla/mux in a number of other CloudFoundry projects

[#184006000](https://www.pivotaltracker.com/story/show/184006000)