fix: [MSSQL] read server_pairs default from config
Although we were exposing a config to allow passing server_pairs
to csb-azure-mssql-fog-run-failover service, it was never used.

CSB complained whenever we tried to create an instance without
specifying server_pairs explicitly.

The name of the property in the config file is very intriguing:
service.csb-azure-mssql-fog-run-failover.provision.defaults.server_pairs
because it seems to suggest that CSB would read some "magic" paths
to determine default values, etc.
But I don't know if this feature was removed or has never worked.