chore: avoid broker creating tf modules for service definitions
The brokerpak spec allows for different ways of specifying template
files. The keywords `template-ref` and `template` cause the broker
to wrap terraform files in a module definition. This is causing extra
issues during when trying to rename providers during execution, as it
requires an extra `tf init` to install the module prior to running the
rename-provider command.
In order to avoid this extra init (which might fail if there are
providers that have been moved/removed), we change all service offerings
to specify templates with the `template-refs` keyword, which means
templates are not wrapped in a module by the broker.
This commit takes care of all old `template-ref` entries. A next commit
will handle `template` entries.

[#186442543](https://www.pivotaltracker.com/story/show/186442543)