chore: avoid broker wrapping bind templates into a module when there is no template
When the `template` keyword is used in the yaml spec, the broker wraps
the resulting combined terraform into a module. This causes issues when
we want to perform provider renames as it needs an init prior to the
renaming, which can fail. In cases where there is terraform code for
binding, specifying files with `template_refs` instead is enough.
This commit tackls the scenario when there is no code. The temporary
fix we are doing here is to add a non empty noop.tf file which we
specify through `template_refs`.
A later improvement would be to actually modify the broker core to
avoid executing any terraform if there is no terraform to execute.

[#183286656](https://www.pivotaltracker.com/story/show/183286656)