chore: avoid broker creating tf modules for service definitions (#621)
* chore: avoid broker creating tf modules for service definitions

The brokerpak spec allows for different ways of specifying template
files. The keywords `template-ref` and `template` cause the broker
to wrap terraform files in a module definition. This is causing extra
issues during when trying to rename providers during execution, as it
requires an extra `tf init` to install the module prior to running the
rename-provider command.
In order to avoid this extra init (which might fail if there are
providers that have been moved/removed), we change all service offerings
to specify templates with the `template-refs` keyword, which means
templates are not wrapped in a module by the broker.
This commit takes care of all old `template-ref` entries. A next commit
will handle `template` entries.

[#186442543](https://www.pivotaltracker.com/story/show/186442543)

* chore: avoid broker wrapping bind templates into a module
when there is no template

When the `template` keyword is used in the yaml spec, the broker wraps
the resulting combined terraform into a module. This causes issues when
we want to perform provider renames as it needs an init prior to the
renaming, which can fail. In cases where there is terraform code for
binding, specifying files with `template_refs` instead is enough.
This commit tackls the scenario when there is no code. The temporary
fix we are doing here is to add a non empty noop.tf file which we
specify through `template_refs`.
A later improvement would be to actually modify the broker core to
avoid executing any terraform if there is no terraform to execute.

[#183286656](https://www.pivotaltracker.com/story/show/183286656)

* chore: set up default env vars to allow tf upgrades and providers
renames

[#183286656](https://www.pivotaltracker.com/story/show/183286656)