chore: bump Azure Terraform provider to V3
Now that we no longer support subsume, it's possible to update the Azure
Terraform provider to V3. Previously this was problematic as running an
update/upgrade on service instances that supported subsume would try to
run `terraform output` which failed as the schema used by the Azure
provider V2/V3 has changed. With subsume no longer supported, we don't
need to run `terraform output` (trigged by `tf_attribute`) and we don't
see the error any more.

[#183376379](https://www.pivotaltracker.com/story/show/183376379)