feat: restrict upgrade to v1.6 only
In order to reduce the size of the brokerpak, obsolete Terraform
providers have been removed. As a result, to get to this version
through upgrading, you must upgrade to v1.6 first and upgrade
all service instances to v1.6. This can be done via "cf upgrade-service",
or for bulk upgrades consider the plugin: https://github.com/cloudfoundry/upgrade-all-services-cli-plugin
You can then upgrade to this version, and again upgrade all the service instances.

[#186588913](https://www.pivotaltracker.com/story/show/186588913)