fix: mongodb the field 'unique' is not valid for _id
[#186981808](https://www.pivotaltracker.com/story/show/186981808)

We started to see some errors related to 'unique' field being
invalid for _id key. There is very little clarity online about
the issue. More so, if we focus on Azure CosmoDB for MongoDB.

After some trial and error, it seems that mongodb instances
created in location 'westus2' have started imposing this, while
instances created in location 'westus' work without issues with
the same configuration.

Changes in this commit is an attempt to cover our backs and be
prepared in case Azure imposes this validation on all locations
in the near future.

These changes should be backward compatible without any user
intervention. Both, from the point of view of existing plans,
and from the point of view of existing instances.
HOWEVER, we may want to introduce an upgrade test to ensure
edge cases with default values work as expected.