fix: spring-music-mongo-db-alternate-rg example-test
[#186981808](https://www.pivotaltracker.com/story/show/186981808)

Previous commit allowed customizing the list of unique_indexes
as well as (non-unique) indexes. This not only allows adding
custom unique and non-unique indexes, which was not possible
before.

It also allows to workaround the 'unique' is not valid for _id
index specification issue. Since we can now decide to not
include _id key in the list of unique_indexes when needed.