fix: mongodb the field 'unique' is not valid for _id (#676)
* fix: mongodb the field 'unique' is not valid for _id

[#186981808](https://www.pivotaltracker.com/story/show/186981808)

We started to see some errors related to 'unique' field being
invalid for _id key. There is very little clarity online about
the issue. More so, if we focus on Azure CosmoDB for MongoDB.

After some trial and error, it seems that mongodb instances
created in location 'westus2' have started imposing this, while
instances created in location 'westus' work without issues with
the same configuration.

Changes in this commit is an attempt to cover our backs and be
prepared in case Azure imposes this validation on all locations
in the near future.

These changes should be backward compatible without any user
intervention. Both, from the point of view of existing plans,
and from the point of view of existing instances.
HOWEVER, we may want to introduce an upgrade test to ensure
edge cases with default values work as expected.

* fix: spring-music-mongo-db-alternate-rg example-test

[#186981808](https://www.pivotaltracker.com/story/show/186981808)

Previous commit allowed customizing the list of unique_indexes
as well as (non-unique) indexes. This not only allows adding
custom unique and non-unique indexes, which was not possible
before.

It also allows to workaround the 'unique' is not valid for _id
index specification issue. Since we can now decide to not
include _id key in the list of unique_indexes when needed.