chore: workaround issue after bumping azurerm to v3.94.0
[#187172990]

Version 3.94.0 for terraform-provider-azurerm caused issues in
redis acceptance_test/upgrade/update_and_upgrade_redis_test
https://github.com/cloudfoundry/csb-brokerpak-azure/commit/e154a5dcb82cfa251efa0abbd8202d0c0e348a31

We were getting error:
> BadRequest: The following updates can't be processed in one
single request, please send separate request to update them:
'properties.sku.capacity,properties.aadEnableDisable'

This workaround doesn't fixes the problem, and we still don't
fully understand what is the rootcause of the issue.
Because of that, we prefer reverting the following commit
until we can guarantee that customer instances won't fail
or be blocked by this type of errors:
https://github.com/cloudfoundry/csb-brokerpak-azure/commit/e154a5dcb82cfa251efa0abbd8202d0c0e348a31

This commit is just a way of documenting the workaround with
the hope that it can help future us better understand the root
cause of the issue.