chore: revert bump 3.94.0 azure provider (2nd attempt) (#708)
* Revert "build(deps): update manifest dependencies file"

This reverts commit 96880f073d0ee684a2f8e0af8358f1465bff45d9.

* Revert "build(deps): bump terraform providers"

This reverts commit 653ce55da9e0cc1f892327f21c857b307c199551.

* chore: revert bump 3.94.0 of azurerm provider

[#187172990](https://www.pivotaltracker.com/story/show/187172990)

This is the second attempt to revert the bump. The first
attempt failed because we had not added v3.94.0 to the
list of versions to ignore. To learn more about this see:
- https://github.com/cloudfoundry/csb-brokerpak-azure/pull/705
- https://github.com/pivotal/cloud-service-broker-ci/pull/858
- https://github.com/pivotal/cloud-service-broker-ci/pull/859

Version 3.94.0 for terraform-provider-azurerm caused issues in
redis acceptance_test/upgrade/update_and_upgrade_redis_test
https://github.com/cloudfoundry/csb-brokerpak-azure/commit/e154a5dcb82cfa251efa0abbd8202d0c0e348a31

We were getting error:
> BadRequest: The following updates can't be processed in one
single request, please send separate request to update them:
'properties.sku.capacity,properties.aadEnableDisable'

We still don't fully understand what is the root cause of the
issue. Because of that, **we prefer reverting the bump commit**
until we can guarantee that customer instances won't fail or be
blocked by this type of errors.

There is another PR with a workaround for the test, but take
into account that it doesn't fix the problem, and should be
used just as a way to document the issue with the hope that
it can help us better understand the root cause in the future.

https://github.com/cloudfoundry/csb-brokerpak-azure/pull/704