chore: annotate sensitive terraform varaibles
By annotating terraform variables that contain sensitive information, we
ensure that they are not printed out in error messages.

This re-implements #356 which previously had to be reverted due to not
being supported by the version of terraform at the time. Note that we
have not chosen to annotate IDs (e.g. user IDs) as sensitive, but only
secrets/passwords. This is because printing out an ID in an error message
can be very helpful.

[#183474479](https://www.pivotaltracker.com/story/show/183474479)