chore: annotate sensitive terraform outputs
This follows on from #712, in which `var.server_credential_pairs` and
`var.server_credentials` were annotates as sensitive as they contain
passwords. An effect of this is that certain output which do not
actually contain sensitive information are now tainted as sensitive as
they are partially derived from a var that is now annotated as
sensitive. To resolve this, we annotate those outputs as sensitive so as
to avoid Terraform failing.

[#183474479](https://www.pivotaltracker.com/story/show/183474479)