fix: delete hard-code instance name
The instance name is hard-coded in the example test.
This causes a conflict when retrying the test after a failure because the
test creates a resource group infrastructure element before creating the
MSSQL server. This results in an error when attempting to create the same
resource group: "resource group already exists." By removing the hard-coded
instance name, we can generate a random instance name and resource
group name to avoid this problem.

[#184671744](https://www.pivotaltracker.com/story/show/184671744)