feat: Add PostgreSQL Flexible Server offering
The existing PostgreSQL offering only creates single servers which are on the path to retirement and will no longer be available by March 2025.
The new offering creates and manages PostgreSQL Flexible Servers and has the minimal setup to match the Single Server offering.

What is different from the Singel Server offering:
 - Cores property is removed because it is a CSB abstraction that doesn't make sense for flexible servers
 - UseTLS  property is not needed - All flexible severs have TLS enabled by default
 - Removed limitation of Postgres versions (removed enum)

[#187624979](https://www.pivotaltracker.com/story/show/187624979)