feat!: CosmosDB and MongoDB retirement (#834)
The IaC code uses properties that no longer exist in the public API [documentation|https://registry.terraform.io/providers/hashicorp/azurerm/3.116.0/docs/resources/cosmosdb_account.]

These services were marked as deprecated in the last release and are planned to be retired in the next one. Rather than expending resources to fix the tests, we plan to retire the services, delete the tests, and avoid unnecessary expenses.