chore: Fixes broker 'failed to allocate' test errors
We can only have 5 instances of our test broker running. Tests that use the test broker can see 'failed to allocate' errors. This is because in the case of a failed creation of a service in our tests will fail to clean up the test broker, causing other test runs to fail.