feat: replace "azurerm_sql_failover_group" with "azurerm_mssql_failover_group"
The resource "azurerm_sql_failover_group" has been deprecated and is removed in the next
version of the AzureRM provider. In order to prepare for this, we are replacing it with
the new alternative "azurerm_mssql_failover_group". Because OpenTofu will run the delete
of "azurerm_sql_failover_group" in parallel with the creation of "azurerm_mssql_failover_group",
we need to create a resource with a different name so that Azure does not fail with
a name conflict error.

- Whether or not `instance_name` is set, a `-g` will be appended to the name of the failover
  group in Azure
- Because the name of the failover group has changed, existing bindings will cease to work,
  and apps must be rebound and restaged.