feat: replace sql_failover_group with mssql_failover_group
The resource "azurerm_sql_failover_group" has been deprecated and is removed in the next version of the AzureRM provider.
In order to prepare for this, we are replacing it with the new alternative "azurerm_mssql_failover_group".
Because OpenTofu will run the delete of "azurerm_sql_failover_group" in parallel with the creation of "azurerm_mssql_failover_group",
the upgrade of service instances will fail at the first attempt. But it will succeed on a second attempt.
This is not an ideal behavior, but as long as it's expected it's easy to deal with, and alternatives
come with a different set of tradeoffs (principally needing to re-bind service instances, or re-engineering the core broker).

Requires core broker fix: https://github.com/cloudfoundry/cloud-service-broker/pull/1179