chore: consistently use DeferCleanup() (#988)
A Ginkgo DeferCleanup() is run _after_ a test finishes
A Go defer() is run as a function ends

A change was made in commit aec88cbcce2f67a8385dbeb3365fd225f3ed27e3
that changed a defer() into a DeferCleanup(), which had the side-effect
of changing cleanup order, and caused cleanup to fail.

This was resolved by consistently using DeferCleanup() in that test.