chore: consistently use DeferCleanup() (#988)
A Ginkgo DeferCleanup() is run _after_ a test finishes
A Go defer() is run as a function ends

A change was made in commit 587d01db416a3c8d53039645e9a76766442eb953
that changed a defer() into a DeferCleanup(), which had the side-effect
of changing cleanup order, and caused cleanup to fail.

This was resolved by consistently using DeferCleanup() in that test.